add rating and review function; improve the layout

+++ DrMediScanV6/Models/ViewModels/SelectedClinic.cs
@@ -6,7 +6,7 @@
         public int ClinicId { get; set; }
         public string ClinicName { get; set; }
         public DateTime AvailableDate { get; set; }
-        public int AverageRate { get; set; }
+        public double AverageRate { get; set; }
         public bool IfAvailable { get; set; }
     }
 }
